refactor(db): move overview related things from db.go to db/overview.go
